requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
scrapy>=2.11.0
requests>=2.31.0
beautifulsoup4>=4.12.0
openai>=1.0.0
sqlalchemy>=2.0.0
alembic>=1.12.0
redis>=5.0.0
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-dotenv>=1.0.0
httpx>=0.25.0
aiofiles>=23.0.0
jinja2>=3.1.0